requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
sqlparse==0.5.5
tzdata==2025.3
Pillow==12.1.0
matplotlib==3.9.3
reportlab==4.2.5
Faker==33.4.0
django-cors-headers==4.6.0
python-decouple==3.8
cx-Oracle==8.3.0
django-filter==24.3
